requirements.txt
fastapi==0.128.5
uvicorn==0.40.0
SQLAlchemy==2.0.46
asyncpg==0.31.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic==2.12.5
alembic==1.18.3
python-jose[cryptography]
passlib[bcrypt]
python-multipart